# kechow-server/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Kechow API",
        "description": "API documentation for Kechow",
        "contact": {
            "email": "support@kechow.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/restaurants": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get all restaurants",
                "operationId": "b838384b35597707b6c46925fe4f77f1",
                "responses": {
                    "200": {
                        "description": "List of restaurants"
                    }
                }
            },
            "post": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Create a new restaurant",
                "operationId": "02d8bcd4a7ef5fb81279b27e4779c593",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Restaurant created"
                    }
                }
            }
        },
        "/api/restaurants/{id}": {
            "get": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Get a specific restaurant",
                "operationId": "0e3fa956792712098b655ec817d9fea7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Restaurant data"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Update a restaurant",
                "operationId": "ccffcce6b9eddf299a250d126812bcf8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Restaurant updated"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Restaurants"
                ],
                "summary": "Delete a restaurant",
                "operationId": "1e632e1ba0da1dca77e48a8c751384f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    }
                }
            }
        },
        "/api/menu-items": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get all menu items",
                "operationId": "f8f6266a0f281d4a6fd2924521990d27",
                "responses": {
                    "200": {
                        "description": "List of menu items"
                    }
                }
            },
            "post": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Create a new menu item",
                "operationId": "080371a7ab243e7fe3fb2e21af1190a8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "restaurant_id",
                                    "name",
                                    "price"
                                ],
                                "properties": {
                                    "restaurant_id": {
                                        "type": "integer"
                                    },
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Menu item created"
                    }
                }
            }
        },
        "/api/menu-items/{id}": {
            "get": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Get a specific menu item",
                "operationId": "9a846f4be41fcdbe9832b94b39f95bf4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Menu item details"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Update a menu item",
                "operationId": "65192842ac27acc7bbe511a3e3d42215",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "is_available": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Menu item updated"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Menu Items"
                ],
                "summary": "Delete a menu item",
                "operationId": "d28acee26f2a00a051b272f4b60af611",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted successfully"
                    },
                    "404": {
                        "description": "Not found"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "MenuItem": {
                "title": "MenuItem",
                "description": "Menu Item model",
                "required": [
                    "restaurant_id",
                    "name",
                    "price"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "restaurant_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Cheeseburger"
                    },
                    "description": {
                        "type": "string",
                        "example": "Juicy beef patty with cheese",
                        "nullable": true
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 8.99
                    },
                    "image_url": {
                        "type": "string",
                        "example": "https://example.com/image.jpg",
                        "nullable": true
                    },
                    "is_available": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "MenuItem"
                }
            },
            "Restaurant": {
                "title": "Restaurant",
                "description": "Restaurant model",
                "required": [
                    "name",
                    "address",
                    "city"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Burger Palace"
                    },
                    "description": {
                        "type": "string",
                        "example": "Best burgers in town"
                    },
                    "address": {
                        "type": "string",
                        "example": "123 Main St"
                    },
                    "city": {
                        "type": "string",
                        "example": "New York"
                    },
                    "state": {
                        "type": "string",
                        "example": "NY"
                    },
                    "zip_code": {
                        "type": "string",
                        "example": "10001"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+1-212-555-1234"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "info@burgerpalace.com"
                    },
                    "website": {
                        "type": "string",
                        "example": "https://burgerpalace.com"
                    },
                    "opening_time": {
                        "type": "string",
                        "example": "10:00"
                    },
                    "closing_time": {
                        "type": "string",
                        "example": "22:00"
                    },
                    "logo_url": {
                        "type": "string",
                        "example": "https://example.com/logo.png"
                    },
                    "is_active": {
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Restaurant"
                }
            },
            "User": {
                "title": "User",
                "description": "User model",
                "required": [
                    "name",
                    "email",
                    "password"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "johndoe@example.com"
                    },
                    "email_verified_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        }
    },
    "tags": [
        {
            "name": "Restaurants",
            "description": "Restaurant operations"
        },
        {
            "name": "Menu Items",
            "description": "Menu Item operations"
        }
    ]
}